chore: Remove unnecesary functions
