+++ src/libs/firebase/store.ts
@@ -2,41 +2,14 @@ import {
   addDoc,
   collection,
   doc,
-  getDocs,
-  query,
   updateDoc,
 } from "firebase/firestore";
 import type { CollectionReference } from "firebase/firestore";
 import { getDownloadURL } from "firebase/storage";
 import { db } from "./init";
 import { uploadImage } from "./storage";
+import type { CreateStoreCardListItem, CreateStoreMenuListItem, StoreCardListItem, StoreMenuListItem } from "./types";
 
-export type CreateStoreMenuListItem = {
-  name: string;
-  date: string;
-  title: string;
-};
-
-export type StoreMenuListItem = {
-  name: string;
-  date: string;
-  id: string;
-};
-
-export type StoreMenuList = StoreMenuListItem[];
-
-export type StoreCardListItem = {
-  text: string;
-  imageRef?: string;
-  createdAt: Date;
-};
-
-export type CreateStoreCardListItem = {
-  text: string;
-  imageFile?: File;
-};
-
-export type StoreCardList = StoreCardListItem[];
 
 export const getMenuListRef = (uuid: string): CollectionReference =>
   collection(db, "Users", uuid, "List");
@@ -61,24 +34,6 @@ export const createMenuListItem = async (
   };
 };
 
-// List全取得
-export const getMenuList = async (uuid: string): Promise<StoreMenuList> => {
-  const ref = query(getMenuListRef(uuid));
-  const list: StoreMenuList = [];
-  // TODO:pushの計算量的に変えるかもしれない
-  getDocs(ref).then((snapshot) =>
-    snapshot.forEach((docs) =>
-      list.push({
-        name: docs.data().name,
-        date: docs.data().date,
-        id: docs.data().id,
-      }),
-    ),
-  );
-
-  return list;
-};
-
 export const createCardListItem = async (
   uuid: string,
   listId: string,
@@ -98,24 +53,3 @@ export const createCardListItem = async (
   return storeData;
 };
 
-export const getCardList = async (
-  uuid: string,
-  listId: string,
-): Promise<StoreCardList> => {
-  const ref = query(getCardListRef(uuid, listId));
-  const list: StoreCardList = [];
-  // TODO:pushの計算量的に変えるかもしれない
-  await getDocs(ref).then((snapshot) =>
-    snapshot.forEach((docs) =>
-      list.push({
-        text: docs.data().text,
-        imageRef: docs.data().date,
-        createdAt: docs.data().createdAt.toDate(),
-      }),
-    ),
-  );
-
-  list.sort((a, b) => (a.createdAt > b.createdAt ? 1 : -1));
-
-  return list;
-};
